requirements.txt
# Core
numpy>=1.24
pandas>=2.0
scipy>=1.10
scikit-learn>=1.3
xarray>=2023.6
netCDF4>=1.6

# Geospatial
rasterio>=1.3
geopandas>=0.14
shapely>=2.0
pyproj>=3.6
osmnx>=1.6
pystac-client>=0.7

# Satellite data
sentinelhub>=3.9
copernicusmarine>=1.0
cdsapi>=0.6

# Deep learning
torch>=2.0
segmentation-models-pytorch>=0.3

# Visualization & reporting
matplotlib>=3.7
fpdf2>=2.7
rich>=13.0

# Utilities
python-dotenv>=1.0
pyyaml>=6.0
tqdm>=4.65

# Server & API
django>=5.0
djangorestframework>=3.14
django-cors-headers>=4.3
pymysql>=1.1.0
